+++ packages/lumx-react/src/components/tabs/Tabs.stories.tsx
@@ -1,4 +1,4 @@
-import { Dialog, Tab, TabList, TabPanel, TabProvider } from '@lumx/react';
+import { Button, Dialog, Tab, TabList, TabPanel, TabProvider } from '@lumx/react';
 import { number } from '@storybook/addon-knobs';
 import get from 'lodash/get';
 import times from 'lodash/times';
@@ -82,23 +82,36 @@ export const DisabledTab = ({ theme }: any) => (
 );
 
 /* Display tabs far from their tab panels. */
-export const SplitTabListAndTabPanels = ({ theme }: any) => (
-    <TabProvider isLazy={false}>
-        <Dialog isOpen forceFooterDivider>
-            <TabPanel className="lumx-spacing-padding-huge">Tab 1 content</TabPanel>
-            <TabPanel className="lumx-spacing-padding-huge">Tab 2 content</TabPanel>
-            <TabPanel className="lumx-spacing-padding-huge">Tab 3 content</TabPanel>
+export const SplitTabListAndTabPanels = ({ theme }: any) => {
+    const [isOpen, setOpen] = useState(true);
+    const [activeTabIndex, onChange] = useState(1);
 
-            <footer>
-                <TabList theme={theme} aria-label="Tab list">
-                    <Tab label="Tab 1" />
-                    <Tab label="Tab 2" isDisabled />
-                    <Tab label="Tab 3" />
-                </TabList>
-            </footer>
-        </Dialog>
-    </TabProvider>
-);
+    return (
+        <TabProvider activeTabIndex={activeTabIndex} onChange={onChange} isLazy={false}>
+            <Button
+                onClick={() => {
+                    setOpen(!isOpen);
+                    onChange(1);
+                }}
+            >
+                Open dialog with tabs in footer
+            </Button>
+            <Dialog isOpen={isOpen} forceFooterDivider onClose={() => setOpen(false)}>
+                <TabPanel className="lumx-spacing-padding-huge">Tab 1 content</TabPanel>
+                <TabPanel className="lumx-spacing-padding-huge">Tab 2 content</TabPanel>
+                <TabPanel className="lumx-spacing-padding-huge">Tab 3 content</TabPanel>
+
+                <footer>
+                    <TabList theme={theme} aria-label="Tab list">
+                        <Tab label="Tab 1" />
+                        <Tab label="Tab 2" />
+                        <Tab label="Tab 3" />
+                    </TabList>
+                </footer>
+            </Dialog>
+        </TabProvider>
+    );
+};
 
 /* Dynamically generate tabs. */
 export const DynamicTabs = ({ theme }: any) => {