+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fixed `TabProvider` setting the `activeTabIndex` to `-1` on unmount.
 ## [1.0.11][] - 2021-02-26
 
 ### Added