+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fixed tabs and progress tracker active index synchronization.
+
 ## [1.0.11][] - 2021-02-26
 
 ### Added